pixel.h and primitives.h, int replaced with intptr_t

+++ source/common/primitives.h
@@ -218,7 +218,7 @@ typedef void (*copy_sp_t)(pixel *dst, intptr_t dstStride, int16_t *src, intptr_t
 typedef void (*copy_ps_t)(int16_t *dst, intptr_t dstStride, pixel *src, intptr_t srcStride);
 
 typedef void (*pixel_sub_ps_t)(int16_t *dst, intptr_t dstride, pixel *src0, pixel *src1, intptr_t sstride0, intptr_t sstride1);
-typedef void (*pixel_add_ps_t)(pixel *a, int dstride, pixel *b0, int16_t *b1, int sstride0, int sstride1);
+typedef void (*pixel_add_ps_t)(pixel *a, intptr_t dstride, pixel *b0, int16_t *b1, intptr_t sstride0, intptr_t sstride1);
 
 /* Define a structure containing function pointers to optimized encoder
  * primitives.  Each pointer can reference either an assembly routine,
